+++ tests/build-lib/RunTest-Unix.bash
@@ -16,12 +16,18 @@ then
 fi
 
 mkdir build-lib-${MODE} && cd build-lib-${MODE}
-cmake -DLIB_BUILD_MODE=${MODE} -G "${CMAKE_GENERATOR}" ${TEST_PROJECT_DIR}
-cmake --build .
+cmake -DLIB_BUILD_MODE=${MODE} -G "${CMAKE_GENERATOR}" -DCMAKE_BUILD_TYPE=Debug ${TEST_PROJECT_DIR}
+cmake --build . --config Debug
 
 
 echo "Checking results..."
 source ${CHECK_RESULTS_LIB}
+
 platform=$(print_build_platform_name)
+if [ $? -ne 0 ]
+then
+  exit 1
+fi
+
 EXPECTED_ARTIFACTS_FILE=${TEST_PROJECT_DIR}/../testconfigs/${MODE}/ExpectedArtifacts.${platform}
 check_build_dir_artifacts ${EXPECTED_ARTIFACTS_FILE}