+++ tests/thirdparty-lib/RunTest-Unix.bash
@@ -4,6 +4,7 @@ set -e
 
 TEST_PROJECT_DIR=$1
 CHECK_RESULTS_LIB=$2
+CMAKE_GENERATOR=$3
 
 echo "Running in $(pwd) with test project dir ${TEST_PROJECT_DIR}"
 
@@ -14,20 +15,17 @@ then
 fi
 
 mkdir thirdparty-lib && cd thirdparty-lib
-cmake ${TEST_PROJECT_DIR}
-cmake --build .
-
-host_os=$(uname)
-if [[ ${host_os} == CYGWIN* ]]
-then
-  host_os=Cygwin
-elif [[ ${host_os} == MSYS* ]]
-then
-  host_os=MSys
-fi
-
-EXPECTED_ARTIFACTS_FILE=${TEST_PROJECT_DIR}/../testconfigs/ExpectedArtifacts.${host_os}
-source ${CHECK_RESULTS_LIB}
+cmake -G "${CMAKE_GENERATOR}" -DCMAKE_BUILD_TYPE=Debug ${TEST_PROJECT_DIR}
+cmake --build . --config Debug
 
 echo "Checking results..."
+source ${CHECK_RESULTS_LIB}
+
+platform=$(print_build_platform_name)
+if [ $? -ne 0 ]
+then
+  exit 1
+fi
+
+EXPECTED_ARTIFACTS_FILE=${TEST_PROJECT_DIR}/../testconfigs/ExpectedArtifacts.${platform}
 check_build_dir_artifacts ${EXPECTED_ARTIFACTS_FILE}